sdk/tests/conformance/typedarrays/data-view-crash.html

<!--
Copyright (c) 2012 The Chromium Authors. All rights reserved.
Use of this source code is governed by a BSD-style license that can be
found in the LICENSE file.
 -->
<html>
<head>
<link rel="stylesheet" href="../../resources/js-test-style.css"/>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>

<script>
description("Test that DataView does not crash with bad offset or length.");

var array = new Uint8Array([164, 112, 157, 63]);
var view;
shouldThrow("view = new DataView(array.buffer, -4500000000)");
shouldThrow("view = new DataView(array.buffer, -4500000000, 4500000000)");
var value = view ? view.getFloat32(0, true) : 0;

successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
